hand_analysis_service/requirements.txt
fastapi-cache2[redis]
fastapi[all]
uvicorn[standard]
motor